MOve large example into separate file and test independently.
